client/requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.7
decorator==4.4.1
fabric==2.4.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
invoke==1.2.0
ipython==7.12.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
more-itertools==7.0.0
numpy==1.16.3
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
paramiko==2.4.2
parso==0.6.1
pdfrw==0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.0.0
pluggy==0.9.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.0
PySide2==5.12.3
pytest==4.4.1
python-dateutil==2.8.0
PyYAML==5.3
reportlab==3.5.20
requests==2.22.0
scipy==1.2.1
shiboken2==5.12.3
six==1.12.0
svg.path==2.2
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
